Move LLVMNG stuff into their own flavours.
This is *not* a user flavour. This is a custom flavour. I don't want to loose this with the subtree/submodule switch every time.

Unearthed another issue: as we require to know all `knownPackages` before hand, we can't simply aggregate the packages from the flavours. That would be really nice to have. The only solution I see is to drop the generation of explicit rules in favour of generic rules, from which we then figure out the package based on the matched path. I'm not sure about the performance impact here though.