Make sure we pass `-DNOSMP` to rts.
This does it in a double fashion. We configure `rts` with `-smp`, which should cause cabal to expose the cpp arg to `-DNOSMP`, but for good measure add it explicitly as well.

Eventually we should migrate all those flags into the `.cabal` file. However the way dependent flags are going to be challenging as cabal has no concept of building `ways` of a library.